Attempt at using computational graphs for SGD
An attempt to build the neural network and optimize it by using
a computational graph structure. Each operation is a node and each
operand is a parameter. The node updates the gradients of its inputs
using the gradient of its output. Works need to be done to debug the
code.